pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "torchradio"
version = "0.0.3"
authors = [
  { name="Daniel Gibbons", email="daniel.gibbons@dewc.com" },
]
description = "A PyTorch differentiable RF simulation environment"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "scikit-commpy",
    "torch",
]

[project.optional-dependencies]
notebooks = [
    "chart_studio",
    "ipykernel",
    "ipywidgets",
    "kaleido",
    "matplotlib",
    "pandas",
    "tqdm",
    "widgetsnbextension",
]

docs = [
    "mkdocs",
    "mkdocs-jupyter",
    "mkdocs-material",
    "mkdocstrings[python]",
    "mknotebooks",
    "pymdown-extensions",
]

dev = [
    "coverage",
    "mypy",
    "pre-commit",
    "pytest",
    "ruff",
    "torchradio[docs]",
    "torchradio[notebooks]",
    "types-tqdm",
]

[project.urls]
"Homepage" = "https://github.com/dewcservices/torchradio"
"Bug Tracker" = "https://github.com/dewcservices/torchradio/issues"

[tool.mypy]
disable_error_code = ["call-overload"]

[tool.ruff]
lint.select = ["ALL"]

lint.ignore = [
    "ANN002",
    "ANN003",
    "ANN101",
    "ANN102",
    "COM812",
    "D100",
    "D104",
    "D203",
    "D211",
    "D213",
    "D407",
    "E501",
    "G004",
    "INP001",
    "ISC001",
    "PGH003",
    "T201",
    "T203",
    "TCH001",
    "TCH002",
    "TCH003",
]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,examples}/*" = ["S101"]
